Test that a YAML file given as a Cue file fails
This test is to check that invalid values files aren't silently
ignored. This is difficult to be sure of, because lots of YAML forms
are accidentally parseable as Cue without representing the value you'd
expect. For example,

```
values:
  level:
  - 1
```

is

    { values: { level: [1] } }

when parsed as YAML, and

    { values: { level: -1 } }

when parsed as Cue. But, we can at least check that genuinely bad
syntax will not be passed over.

Signed-off-by: Michael Bridgen <mikeb@squaremobius.net>