Test case for using YAML files as values
This test case checks that you can supply a YAML file to `timoni
build`, and it will use the data therein as values like those botained
from a .cue file.

THere is a subtlety: the values file used in the test is also valid
Cue, but does not represent the same value as when treated as
YAML*. If it were invalid as Cue, that would not test the possibility
of executing without errors, but getting the answer wrong.

*The domain field is part of the `values` struct in YAML, but at the
top level in Cue (where indentation is not significant).

Signed-off-by: Michael Bridgen <mikeb@squaremobius.net>