Implement YAML/JSON values file conversion
This commit gives the simple and adequate implementation for loading
JSON and YAML files of using the packages under encoding/ directly
(which are advertised as being for converting values to and from Cue).

The code supporting the command `cue import`, which is an analogue for
what's needed here, has many more concerns to deal with -- other
formats, schemas, packages. It does a lot of work, via many
abstractions (mostly unexported or internal), to interpret what it's
given. We have no need of the complication, since we need only to
parse each values file in the format it purports to use, and convert
it to a Cue value.

Signed-off-by: Michael Bridgen <mikeb@squaremobius.net>