Pass values as []byte rather than files in $TMP
This rewrites the values file handling such that it avoids writing a
temp file when converting from JSON and YAML. This means values files
get passed to the engine as []byte -- if it's a cue file, it's just
the contents of the file.

Reading a values file from stdin (by supplying `-f -`) is handled in
the convertToCue helper too, since that func takes the paths and
returns the bytes. And since it's now bytes, saveReaderToFile
(previously used so what's read on stdin can be supplied as a file),
isn't needed by apply, so I moved it.

This extends the support for JSON and YAML values files to `timoni
apply` as well.

Signed-off-by: Michael Bridgen <mikeb@squaremobius.net>