Add output flag
The output flag will take a file name and read the file line by line,
write each line to a new temporary file. If the file is a markdown file
we are looking for a table which will be replaced with the new table
generated from the module configuration data. If a table is not found or
the file is not markdown then the generated table is appended to the end
of the file. The temporary file is then used to replace the existing
file.

Add logic to look for specific markdown header and only work with the table under that
Better replace for json structure
Fix formatting for types with pipe values
Use string replace on correct variable

Signed-off-by: Luke Mallon (Nalum) <luke@mallon.ie>