Update to cuelang.org/go@v0.9.1
In so doing, set cue/load.Config.AcceptLegacyModules when loading
modules, to allow pre-existing Timoni modules (which don't have a
language.version field) to be loaded. For more information see
https://cuelang.org/issue/3219

Note that https://cuelang.org/issue/3222 tracks making a change in a
later CUE release that will "fix" the language.version implied by
AcceptLegacyModules. This is not significant right now, because there
are not yet any language changes guarded by language.version, but it
will become significant when such changes are introduced.