Fix CLI autocompletion on ZSH when installed with Homebrew
Prior to this, when installing Timoni using Homebrew on a Mac with ZSH,
CLI autocompletion would not work.

After this change, CLI autocompletion works after installation or
reinstallation via Homebrew.

The problem is that the completion function that Homebrew wrote to
`/opt/homebrew/share/zsh/site-functions/` did not start with an
underscore, `_`.

By convention, completion function names (and the files containing them)
should start with an underscore followed by the name of the command they
provide completions for. When ZSH's completion system is initialized
(usually through the compinit function), it automatically looks for
files in the fpath directories that start with an underscore.